Update bindings to use TargetModel
