+++ python/AIERTXModule.cpp
@@ -8,7 +8,10 @@
 //
 //===----------------------------------------------------------------------===//
 
+#include "aie-c/TargetModel.h"
 #include "aie-c/Translation.h"
+
+#include "PyTypes.h"
 
 #include <pybind11/pybind11.h>
 #include <pybind11/pytypes.h>
@@ -20,8 +23,8 @@ using namespace py::literals;
 
 class PyAIERTXControl {
 public:
-  PyAIERTXControl(size_t partitionNumCols)
-      : ctl(getAieRtxControl(partitionNumCols)) {}
+  PyAIERTXControl(AieTargetModel targetModel)
+      : ctl(getAieRtxControl(targetModel)) {}
 
   ~PyAIERTXControl() { freeAieRtxControl(ctl); }
 
@@ -31,7 +34,7 @@ public:
 PYBIND11_MODULE(_aiertx, m) {
 
   py::class_<PyAIERTXControl>(m, "AIERTXControl", py::module_local())
-      .def(py::init<size_t>(), "partition_num_cols"_a)
+      .def(py::init<PyAieTargetModel>(), "target_model"_a)
       .def("start_transaction",
            [](PyAIERTXControl &self) { aieRtxStartTransaction(self.ctl); })
       .def("export_serialized_transaction",
