+++ test/python/aiertx_bindings.py
@@ -9,12 +9,13 @@
 from aie.aiertx import AIERTXControl
 from util import construct_and_print_module
 from aie.dialects.aiex import DDR_AIE_ADDR_OFFSET
-
+from aie.dialects.aie import AIEDevice, get_target_model
 
 # CHECK-LABEL: simple
 @construct_and_print_module
 def simple(module):
-    ctl = AIERTXControl(4)
+    tm = get_target_model(AIEDevice.npu1_4col)
+    ctl = AIERTXControl(tm)
     ctl.start_transaction()
     ctl.dma_update_bd_addr(0, 0, DDR_AIE_ADDR_OFFSET, 0)
     ctl.export_serialized_transaction()
